+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 * Add `CodeQL` check in CI
+* Add parsing `Presentation#slide_master`
 
 ### Changes